Closes #418

+++ PKGBUILD
@@ -14,37 +14,70 @@ pkgver() {
 }
 
 build() {
+  local compiler=other
+  local do_pgo=no
+
   # makepkg uses -O2 by default, but we *really* want -O3
-  # -Wno-missing-profile since we're not exercising everything when doing PGO builds
-  export CFLAGS+=" -O3 -Wno-missing-profile"
+  CFLAGS+=" -O3"
+
+  # Figure out which compiler we're using, and whether or not to do PGO
+  case $(${CC-cc} --version) in
+    *GCC*)
+      compiler=gcc
+      do_pgo=yes
+      ;;
+
+    *clang*)
+      compiler=clang
+
+      # We need llvm to be able to manage the profiling data
+      if command -v llvm-profdata > /dev/null; then
+        do_pgo=yes
+
+        # Meson adds -fprofile-correction, which Clang doesn't
+        # understand
+        CFLAGS+=" -Wno-ignored-optimization-argument"
+      fi
+      ;;
+  esac
 
   meson --prefix=/usr --buildtype=release --wrap-mode=nofallback -Db_lto=true ..
 
-  find -name "*.gcda" -delete
-  meson configure -Db_pgo=generate
-  ninja
+  if [[ ${do_pgo} == yes ]]; then
+    find -name "*.gcda" -delete
+    meson configure -Db_pgo=generate
+    ninja
 
-  script_options="--scroll --scroll-region --colors-regular --colors-bright --colors-256 --colors-rgb --attr-bold --attr-italic --attr-underline --sixel"
+    local script_options="--scroll --scroll-region --colors-regular --colors-bright --colors-256 --colors-rgb --attr-bold --attr-italic --attr-underline --sixel"
 
-  tmp_file=$(mktemp)
+    tmp_file=$(mktemp)
 
-  if [[ -v WAYLAND_DISPLAY ]]; then
-    ./foot \
-      --config /dev/null \
-      --term=xterm \
-      sh -c "../scripts/generate-alt-random-writes.py ${script_options} ${tmp_file} && cat ${tmp_file}"
-  else
-    ../scripts/generate-alt-random-writes.py \
-      --rows=67 \
-      --cols=135 \
-      ${script_options} \
-      ${tmp_file}
-    ./pgo ${tmp_file} ${tmp_file} ${tmp_file}
+    if [[ -v WAYLAND_DISPLAY ]]; then
+      ./footclient --version
+      ./foot \
+        --config /dev/null \
+        --term=xterm \
+        sh -c "../scripts/generate-alt-random-writes.py ${script_options} ${tmp_file} && cat ${tmp_file}"
+    else
+      ./footclient --version
+      ./foot --version
+      ../scripts/generate-alt-random-writes.py \
+        --rows=67 \
+        --cols=135 \
+        ${script_options} \
+        ${tmp_file}
+      ./pgo ${tmp_file} ${tmp_file} ${tmp_file}
+    fi
+
+    rm "${tmp_file}"
+
+    if [[ ${compiler} == clang ]]; then
+      llvm-profdata merge default_*profraw --output=default.profdata
+    fi
+
+    meson configure -Db_pgo=use
   fi
 
-  rm "${tmp_file}"
-
-  meson configure -Db_pgo=use
   ninja
 }
